wip: removed code execution from validation
